+++ plugin/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.36](https://github.com/Akylas/nativescript-chart/compare/v0.0.35...v0.0.36) (2020-09-27)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+
+
+
+
+## [0.0.35](https://github.com/Akylas/nativescript-chart/compare/v0.0.34...v0.0.35) (2020-09-27)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+
+
+
+
+## [0.0.34](https://github.com/Akylas/nativescript-chart/compare/v0.0.33...v0.0.34) (2020-09-27)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+
+
+
+
+## [0.0.33](https://github.com/Akylas/nativescript-chart/compare/v0.0.32...v0.0.33) (2020-09-27)
+
+**Note:** Version bump only for package @nativescript-community/ui-chart
+
+
+
+
+
 ## [0.0.32](https://github.com/Akylas/nativescript-chart/compare/v0.0.31...v0.0.32) (2020-09-22)
 
 **Note:** Version bump only for package @nativescript-community/ui-chart